+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-2.11 (unreleased)
------------------
+3.0 (unreleased)
+----------------
 
 - Fixed javascript error for undefined tinymce due to typo.  Note that
   on Plone 4 we currently cannot load TinyMCE.